combine google and tavily into search.py
