+++ deeppresenter/mcp.json.example
@@ -51,7 +51,8 @@
             "$PACKAGE_DIR/tools/search.py"
         ],
         "env": {
-            "TAVILY_API_KEY": "your_tavily_api_key_here"
+            "TAVILY_API_KEY": "your_tavily_api_key_here",
+            "SERPAPI_KEY": "your_seriapi_api_key_here"
         },
         "network": true
     },
