+++ docker-compose.yml
@@ -14,7 +14,14 @@ services:
     environment:
       DEEPPRESENTER_HOST_WORKSPACE_BASE: ${PWD}/workspace
       DEEPPRESENTER_WORKSPACE_BASE: /opt/workspace
-      DEEPPRESENTER_LOG_LEVEL: 0 # 0: debug, 10: info, 20: warning
+      DEEPPRESENTER_LOG_LEVEL: 0
+      PLAYWRIGHT_BROWSERS_PATH: /ms-playwright
+      http_proxy:
+      https_proxy:
+      HTTP_PROXY:
+      HTTPS_PROXY:
+      no_proxy: localhost,127.0.0.1,0.0.0.0
+      NO_PROXY: localhost,127.0.0.1,0.0.0.0
     volumes:
       - ./workspace:/opt/workspace
       - ./deeppresenter:/usr/src/pptagent/deeppresenter
